move ContainerRemoteSerializer back to ui
Issue: AAH-766
